Modularize testing classes
